+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 ### Ported from [kamilsk/cilex-service-providers](https://github.com/kamilsk/CilexServiceProviders)
 - classes
   - `OctoLab\Cilex\ServiceProvider\ConfigServiceProvider`
+  - `OctoLab\Cilex\ServiceProvider\DoctrineServiceProvider`
   - `OctoLab\Cilex\ServiceProvider\MonologServiceProvider`